requirements.txt
Flask==3.1.0
numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.0.0
torch==2.4.1
torch==2.5.1
Werkzeug==3.1.3
